+++ src/main/java/net/sf/xenqtt/client/ProgressiveReconnectionStrategy.java
@@ -38,7 +38,7 @@ public final class ProgressiveReconnectionStrategy implements ReconnectionStrate
 	 *            The maximum number of reconnect attempts to make
 	 */
 	ProgressiveReconnectionStrategy(long baseReconnectMillis, int progressiveFactor, int maxNumberOfReconnects) {
-		this.baseReconnectMillis = (Long) XenqttUtil.validateGreaterThan("baseReconnectMillis", baseReconnectMillis, 0);
+		this.baseReconnectMillis = XenqttUtil.validateGreaterThan("baseReconnectMillis", baseReconnectMillis, 0L);
 		this.progressiveFactor = XenqttUtil.validateGreaterThan("progressiveFactor", progressiveFactor, 0);
 		this.maxNumberOfReconnects = XenqttUtil.validateGreaterThanOrEqualTo("maxNumberOfReconnects", maxNumberOfReconnects, 0);
 		currentRetry = new AtomicInteger();
